requirements.txt
flask
beautifulsoup4
configparser
line-bot-sdk
requests
